Unify health component to event-driven version
