--- UnityGame/Assets/Scripts/Gameplay/Health.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c6f0499acdeb4b739eff660774aac1cd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
-  assetBundleName:
-  assetBundleVariant:
